+++ src/scenes/Home/index.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+class Home extends React.Component {
+  render() {
+    return (
+      <div className="container">
+        <div className="row pt-5">
+          <div className="col-12 col-lg-6 offset-lg-3">
+            <h1 className="text-center">The Lilac Foundation</h1>
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Home;