+++ changelog.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Webpack configuration
 - Babel configuraiton
 - Eslint configuraiton
+
+- SASS folder structure
+- React base component